Proposed new subject for the FAQ
Simple walk-through of how to use the package. I used some time before I found this example in the documentation under eudract_convert(). This is just a nice help for the not-so-technical-or-native-programmer type of use. I am an MD, and very happy to use the tool.
This suggestion could be expanded with Stata and SAS examples.